Finalizacion de validaciones

+++ AsociacionCordobesaDeNatacion/AsociacionCordobesaDeNatacion/AsociacionCordobesaDeNatacion/Formularios/ABM-Clubes.cs
@@ -1,20 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using AsociacionCordobesaDeNatacion.Clases;
-
+using AsociacionCordobesaDeNatacion.Utils;
 namespace AsociacionCordobesaDeNatacion.Formularios
 {
     public partial class ABM_Clubes : Form
     {
         Clubes clubes = new Clubes();
-
+        
         public ABM_Clubes()
         {
             InitializeComponent();
@@ -22,10 +17,7 @@ namespace AsociacionCordobesaDeNatacion.Formularios
 
         private void btn_buscar_Click(object sender, EventArgs e)
         {
-
-
-
-            if (this.txt_cod_club.Text == "")
+            if (txt_cod_club.Text =="")
             {
 
                 MessageBox.Show("El codigo no está cargado");
@@ -58,25 +50,40 @@ namespace AsociacionCordobesaDeNatacion.Formularios
         private void cmd_grabar_Click(object sender, EventArgs e)
         {
 
-            clubes.calle_club = this.txt_calle.Text;
-            clubes.cod_club = this.txt_cod_club.Text;
-            clubes.numero_club = this.txt_numero.Text;
-            clubes.nombre_club = this.txt_nombre_club.Text;
-            this.clubes.grabar_club();
-            MessageBox.Show("La grabación fue correcta");
-            this.blanquear_objetos();
-
+            List<TextBox> array = crearArray();
+            if (Utils.FormValidator.validacionesDeTextosVacios(array))
+            {
+                clubes.calle_club = this.txt_calle.Text;
+                clubes.cod_club = this.txt_cod_club.Text;
+                clubes.numero_club = this.txt_numero.Text;
+                clubes.nombre_club = this.txt_nombre_club.Text;
+                this.clubes.grabar_club();
+                MessageBox.Show("La grabación fue correcta");
+                this.blanquear_objetos();
+            }else
+            {
+                MessageBox.Show("Falta ingresar algun dato");
+            }
         }
 
         private void cmd_actualizar_Click(object sender, EventArgs e)
         {
-            clubes.calle_club = this.txt_calle.Text;
-            clubes.cod_club = this.txt_cod_club.Text;
-            clubes.numero_club = this.txt_numero.Text;
-            clubes.nombre_club = this.txt_nombre_club.Text;
-            this.clubes.modificar_club(this.txt_cod_club.Text);
-            MessageBox.Show("La modificacion fue correcta");
-            this.blanquear_objetos();
+           List<TextBox> array = crearArray();
+            if (Utils.FormValidator.validacionesDeTextosVacios(array))
+            {
+                clubes.calle_club = this.txt_calle.Text;
+                clubes.cod_club = this.txt_cod_club.Text;
+                clubes.numero_club = this.txt_numero.Text;
+                clubes.nombre_club = this.txt_nombre_club.Text;
+                this.clubes.modificar_club(this.txt_cod_club.Text);
+                MessageBox.Show("La modificacion fue correcta");
+                this.blanquear_objetos();
+            }
+            else
+            {
+                MessageBox.Show("Falta ingresar algun dato");
+            }
+            
         }
 
         private void blanquear_objetos()
@@ -89,57 +96,30 @@ namespace AsociacionCordobesaDeNatacion.Formularios
 
         private void btn_eliminar_Click(object sender, EventArgs e)
         {
-            clubes.cod_club = this.txt_cod_club.Text;
-            DialogResult dialogResult = MessageBox.Show("está seguro que desea eliminar el club? Con codigo:" + clubes.cod_club
-                , "importante", MessageBoxButtons.YesNo);
-            if (dialogResult == DialogResult.Yes)
+            if (txt_cod_club.Text != "")
             {
-                clubes.eliminar_club();
-            }
-            else if (dialogResult == DialogResult.No)
-            {
-                dialogResult = DialogResult.Cancel;
-            }
-        }
-        private void restriccionDeLetras_KeyPress(object sender, KeyPressEventArgs e)
-        {
-            if (Char.IsDigit(e.KeyChar))
-            {
-                e.Handled = false;
-            }
-            else if (Char.IsControl(e.KeyChar))
-            {
-                e.Handled = false;
+                clubes.cod_club = this.txt_cod_club.Text;
+                DialogResult dialogResult = MessageBox.Show("está seguro que desea eliminar el club? Con codigo:" + clubes.cod_club
+                    , "importante", MessageBoxButtons.YesNo);
+                if (dialogResult == DialogResult.Yes)
+                {
+                    clubes.eliminar_club();
+                }
+                else if (dialogResult == DialogResult.No)
+                {
+                    dialogResult = DialogResult.Cancel;
+                }
             }
             else
             {
-                e.Handled = true;
+                MessageBox.Show("No ingreso un codigo de club");
             }
         }
-
-        private void restriccionDeNumeros_KeyPress(object sender, KeyPressEventArgs e)
-        {
-            if (Char.IsLetter(e.KeyChar))
-            {
-                e.Handled = false;
-            }
-            else if (Char.IsControl(e.KeyChar))
-            {
-                e.Handled = false;
-            }
-            else if (Char.IsSeparator(e.KeyChar))
-            {
-                e.Handled = false;
-            }
-            else
-            {
-                e.Handled = true;
-            }
-        }
+        
 
         private void txt_cod_club_KeyPress(object sender, KeyPressEventArgs e)
         {
-            restriccionDeLetras_KeyPress(sender,e);
+            Utils.FormValidator.restriccionDeLetras_KeyPress(sender,e);
             
             //no se encuentra en buen funcionamiento
             btn_buscar.Enabled = true;
@@ -147,15 +127,19 @@ namespace AsociacionCordobesaDeNatacion.Formularios
 
         private void txt_numero_KeyPress(object sender, KeyPressEventArgs e)
         {
-            restriccionDeLetras_KeyPress(sender, e);
+            Utils.FormValidator.restriccionDeLetras_KeyPress(sender, e);
         }
 
-        private void txt_nombre_club_KeyPress(object sender, KeyPressEventArgs e)
+        private List<TextBox> crearArray()
         {
-            restriccionDeNumeros_KeyPress(sender, e);
-        }
+            List<TextBox> array = new List<TextBox>();
+            array.Add(this.txt_calle);
+            array.Add(this.txt_cod_club);
+            array.Add(this.txt_numero);
+            array.Add(this.txt_nombre_club);
 
-        
+            return array;
+        }
     }
     
 }
